requirements.txt
matplotlib==3.10.0
numpy==1.26.4
pandas==2.2.3
pip==25.0
PyYAML==6.0.2
scikit-learn==1.6.1
seaborn==0.13.2
optuna==4.4.0
optuna-integration==4.4.0
lightgbm==4.6.0
xgboost==3.0.1
ipywidgets==8.1.5
pytest==8.4.1
pytest-cov==6.0.0
statsmodels==0.14.4